Organization
Organized the files of the project into their separate directories
