+++ Tweeter/accounts/signals.py
@@ -1,7 +1,7 @@
 from django.db.models.signals import post_save
 from django.contrib.auth.models import User
 from django.dispatch import receiver
-from .models import Profile
+from project.models import Profile
 
 
 # Upon new registration, create a profile for the user to be saved
